Replace C-style casts around std::pow calls in DrumSynth with static_cast.

+++ src/core/DrumSynth.cpp
@@ -322,13 +322,13 @@ int DrumSynth::GetDSFileSamples(QString dsfile, int16_t *&wave, int channels, sa
   timestretch *= Fs / 44100.f;
 
   DGain = 1.0f; //leave this here!
-  DGain = (float)std::pow(10.0, 0.05 * GetPrivateProfileFloat(sec,"Level",0,dsfile));
+  DGain = static_cast<float>(std::pow(10.0, 0.05 * GetPrivateProfileFloat(sec,"Level",0,dsfile)));
 
   MasterTune = GetPrivateProfileFloat(sec,"Tuning",0.0,dsfile);
-  MasterTune = (float)std::pow(1.0594631f, MasterTune + mem_tune);
+  MasterTune = static_cast<float>(std::pow(1.0594631f, MasterTune + mem_tune));
   MainFilter = 2 * GetPrivateProfileInt(sec,"Filter",0,dsfile);
   MFres = 0.0101f * GetPrivateProfileFloat(sec,"Resonance",0.0,dsfile);
-  MFres = (float)std::pow(MFres, 0.5f);
+  MFres = static_cast<float>(std::pow(MFres, 0.5f));
 
   HighPass = GetPrivateProfileInt(sec,"HighPass",0,dsfile);
   GetEnv(7, sec, "FilterEnv", dsfile);
@@ -363,7 +363,7 @@ int DrumSynth::GetDSFileSamples(QString dsfile, int16_t *&wave, int channels, sa
   TDroopRate = GetPrivateProfileFloat(sec,"Droop",0.f,dsfile);
   if(TDroopRate>0.f)
   {
-    TDroopRate = (float)std::pow(10.0f, (TDroopRate - 20.0f) / 30.0f);
+    TDroopRate = static_cast<float>(std::pow(10.0f, (TDroopRate - 20.0f) / 30.0f));
     TDroopRate = TDroopRate * -4.f / envData[1][MAX];
     TDroop = 1;
     F2 = F1+((F2-F1)/(1.f-(float)exp(TDroopRate * envData[1][MAX])));
@@ -386,7 +386,7 @@ int DrumSynth::GetDSFileSamples(QString dsfile, int16_t *&wave, int channels, sa
   OW1 = GetPrivateProfileInt(sec,"Wave1",0,dsfile);
   OW2 = GetPrivateProfileInt(sec,"Wave2",0,dsfile);
   OBal2 = (float)GetPrivateProfileInt(sec,"Param",50,dsfile);
-  ODrive = (float)std::pow(OBal2, 3.0f) / (float)std::pow(50.0f, 3.0f);
+  ODrive = static_cast<float>(std::pow(OBal2, 3.0f)) / std::pow(50.0f, 3.0f);
   OBal2 *= 0.01f;
   OBal1 = 1.f - OBal2;
   Ophi1 = Tphi;
@@ -446,8 +446,8 @@ int DrumSynth::GetDSFileSamples(QString dsfile, int16_t *&wave, int channels, sa
   {
     DAtten = DGain * (short)LoudestEnv();
     if(DAtten>32700) clippoint=32700; else clippoint=(short)DAtten;
-    DAtten = (float)std::pow(2.0, 2.0 * GetPrivateProfileInt(sec,"Bits",0,dsfile));
-    DGain = DAtten * DGain * (float)std::pow(10.0, 0.05 * GetPrivateProfileInt(sec,"Clipping",0,dsfile));
+    DAtten = static_cast<float>(std::pow(2.0, 2.0 * GetPrivateProfileInt(sec,"Bits",0,dsfile)));
+    DGain = DAtten * DGain * static_cast<float>(std::pow(10.0, 0.05 * GetPrivateProfileInt(sec,"Clipping",0,dsfile)));
   }
 
   //prepare envelopes
@@ -657,7 +657,7 @@ int DrumSynth::GetDSFileSamples(QString dsfile, int16_t *&wave, int channels, sa
 
         MFtmp = envData[7][ENV];
         if(MFtmp >0.2f)
-          MFfb = 1.001f - (float)std::pow(10.0f, MFtmp - 1);
+          MFfb = 1.001f - static_cast<float>(std::pow(10.0f, MFtmp - 1));
         else
           MFfb = 0.999f - 0.7824f * MFtmp;
 
@@ -675,7 +675,7 @@ int DrumSynth::GetDSFileSamples(QString dsfile, int16_t *&wave, int channels, sa
 
         MFtmp = envData[7][ENV];
         if(MFtmp >0.2f)
-          MFfb = 1.001f - (float)std::pow(10.0f, MFtmp - 1);
+          MFfb = 1.001f - static_cast<float>(std::pow(10.0f, MFtmp - 1));
         else
           MFfb = 0.999f - 0.7824f * MFtmp;
